backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-stubs>=3.16.3",
    "markdown>=3.9",
    "psycopg2-binary>=2.9.9",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.0.0",
]
